Move functions * Bump version to 3.5.3a1 (developoment) * Move functions from jsonod.py to utils.py * Move functions from node.py to printing.py
