+++ src/objdictgen/jsonod.py
@@ -35,6 +35,8 @@ from objdictgen import node as nodelib
 from objdictgen.maps import OD, ODMapping, ODMappingList
 from objdictgen.typing import (TDiffNodes, TIndexEntry, TODJson, TODObjJson,
                                TODObj, TODSubObj, TODSubObjJson, TODValue, TParamEntry, TPath, TProfileMenu)
+from objdictgen.utils import (copy_in_order, exc_amend, maybe_number,
+                              str_to_int)
 
 T = TypeVar('T')
 M = TypeVar('M', bound=Mapping)
@@ -180,53 +182,6 @@ def remove_jsonc(text: str) -> str:
         return match.group(1)
 
     return RE_JSONC.sub(_re_sub, text,)
-
-
-# FIXME: Move to generic utils/funcs?
-def exc_amend(exc: Exception, text: str) -> Exception:
-    """ Helper to prefix text to an exception """
-    args = list(exc.args)
-    if len(args) > 0:
-        args[0] = text + str(args[0])
-    else:
-        args.append(text)
-    exc.args = tuple(args)
-    return exc
-
-
-def str_to_int(string: str|int) -> int:
-    """ Convert string or int to int. Fail if not possible."""
-    i = maybe_number(string)
-    if not isinstance(i, int):
-        raise ValueError(f"Expected integer, got '{string}'")
-    return i
-
-
-def maybe_number(string: str|int) -> int|str:
-    """ Convert string to a number, otherwise pass it through as-is"""
-    if isinstance(string, int):
-        return string
-    s = string.strip()
-    if s.startswith('0x') or s.startswith('-0x'):
-        return int(s.replace('0x', ''), 16)
-    if s.isdigit():
-        return int(string)
-    return string
-
-
-def copy_in_order(d: M, order: Sequence[T]) -> M:
-    """ Remake dict d with keys in order """
-    out = {
-        k: d[k]
-        for k in order
-        if k in d
-    }
-    out.update({
-        k: v
-        for k, v in d.items()
-        if k not in out
-    })
-    return cast(M, out)  # FIXME: For mypy
 
 
 def remove_underscore(d: T) -> T:
